+++ packages/json-comb-core/CHANGELOG.md
@@ -4,6 +4,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [5.0.0] - 2017-12-28
+### Changed
+
+✨✨✨  ASYNC! ✨✨✨
+
+- ✨ All ~~methods~~ exported functions up until now were synchronous. Now all were renamed:
+
+```
+getKeyset      => getKeysetSync
+enforceKeyset  => enforceKeysetSync
+sortAllObjects => sortAllObjectsSync
+noNewKeys      => noNewKeysSync
+findUnused     => findUnusedSync
+```
+
+In their place, async methods are placed. For starters we have async `getKeyset` and `enforceKeyset` - both consume and return promises.
+
+Please update your API's appending "Sync" on the function names, or better, tap async-ones instead.
+
 ## [4.2.0] - 2017-12-12
 ### Changed
 - ✨ Small rebasing, improvements to the setup and proper deps refresh.
@@ -71,3 +90,4 @@ API-wise, there are no changes, but since it's rewrite in ES modules and the arr
 [4.0.0]: https://github.com/codsen/json-comb-core/compare/v3.0.0...v4.0.0
 [4.1.0]: https://github.com/codsen/json-comb-core/compare/v4.0.0...v4.1.0
 [4.2.0]: https://github.com/codsen/json-comb-core/compare/v4.1.0...v4.2.0
+[5.0.0]: https://github.com/codsen/json-comb-core/compare/v4.2.0...v5.0.0